paginas de componentes corrigidas
